refactor: move back blockExistsOnPage
